# Point on line fix