--- a/packages/turf-point-on-line/index.js
+++ b/packages/turf-point-on-line/index.js
@@ -1,8 +1,10 @@
 var distance = require('@turf/distance');
-var point = require('@turf/helpers').point;
+var helpers = require('@turf/helpers');
 var bearing = require('@turf/bearing');
 var destination = require('@turf/destination');
-
+var lineIntersects = require('@turf/line-intersect');
+var point = helpers.point;
+var lineString = helpers.lineString;
 /**
  * Takes a {@link Point} and a {@link LineString} and calculates the closest Point on the LineString.
  *
@@ -71,19 +73,10 @@ module.exports = function (line, pt, units) {
         var direction = bearing(start, stop);
         var perpendicularPt1 = destination(pt, heightDistance, direction + 90, units);
         var perpendicularPt2 = destination(pt, heightDistance, direction - 90, units);
-        var intersect = lineIntersects(
-            perpendicularPt1.geometry.coordinates[0],
-            perpendicularPt1.geometry.coordinates[1],
-            perpendicularPt2.geometry.coordinates[0],
-            perpendicularPt2.geometry.coordinates[1],
-            start.geometry.coordinates[0],
-            start.geometry.coordinates[1],
-            stop.geometry.coordinates[0],
-            stop.geometry.coordinates[1]
-        );
-        var intersectPt;
-        if (intersect) {
-            intersectPt = point(intersect);
+        var intersect = lineIntersects(lineString([perpendicularPt1.geometry.coordinates, perpendicularPt2.geometry.coordinates]), lineString([start.geometry.coordinates, stop.geometry.coordinates]));
+        var intersectPt = null
+        if (intersect.features.length > 0) {
+            intersectPt = intersect.features[0];
             intersectPt.properties.dist = distance(pt, intersectPt, units);
             intersectPt.properties.location = length + distance(start, closestPt, units);
         }
@@ -108,48 +101,3 @@ module.exports = function (line, pt, units) {
 
     return closestPt;
 };
-
-// modified from http://jsfiddle.net/justin_c_rounds/Gd2S2/light/
-function lineIntersects(line1StartX, line1StartY, line1EndX, line1EndY, line2StartX, line2StartY, line2EndX, line2EndY) {
-    // if the lines intersect, the result contains the x and y of the intersection (treating the lines as infinite) and booleans for whether line segment 1 or line segment 2 contain the point
-    var denominator, a, b, numerator1, numerator2;
-    var result = {
-        x: null,
-        y: null,
-        onLine1: false,
-        onLine2: false
-    };
-    denominator = ((line2EndY - line2StartY) * (line1EndX - line1StartX)) - ((line2EndX - line2StartX) * (line1EndY - line1StartY));
-    if (denominator === 0) {
-        if (result.x !== null && result.y !== null) {
-            return result;
-        } else {
-            return false;
-        }
-    }
-    a = line1StartY - line2StartY;
-    b = line1StartX - line2StartX;
-    numerator1 = ((line2EndX - line2StartX) * a) - ((line2EndY - line2StartY) * b);
-    numerator2 = ((line1EndX - line1StartX) * a) - ((line1EndY - line1StartY) * b);
-    a = numerator1 / denominator;
-    b = numerator2 / denominator;
-
-    // if we cast these lines infinitely in both directions, they intersect here:
-    result.x = line1StartX + (a * (line1EndX - line1StartX));
-    result.y = line1StartY + (a * (line1EndY - line1StartY));
-
-    // if line1 is a segment and line2 is infinite, they intersect if:
-    if (a > 0 && a < 1) {
-        result.onLine1 = true;
-    }
-    // if line2 is a segment and line1 is infinite, they intersect if:
-    if (b > 0 && b < 1) {
-        result.onLine2 = true;
-    }
-    // if line1 and line2 are segments, they intersect if both of the above are true
-    if (result.onLine1 && result.onLine2) {
-        return [result.x, result.y];
-    } else {
-        return false;
-    }
-}