- [x] Use a meaningful title for the pull request. Include the name of the package modified.
- [x] Have read [How To Contribute](https://github.com/Turfjs/turf/blob/master/CONTRIBUTING.md#how-to-contribute).
- [x] Run `npm test` at the sub modules where changes have occurred.
- [x] Run `npm run lint` to ensure code style at the turf module level.

Gday @DenisCarriere 

Fix for issue #344 by replacing the custom `lineIntersects` function with the new Turf module.

### Benchmark results
````
turf-point-on-line#simple x 302,308 ops/sec ±0.83% (98 runs sampled)
turf-point-on-line#route1 x 153 ops/sec ±0.43% (87 runs sampled)
turf-point-on-line#route2 x 178 ops/sec ±0.35% (85 runs sampled)
````
Seems a little slow for those route scenarios given how fast the simple scenario is so there is probably some optimisation that could take place I suspect.

Even ran the lint and tests check for once ;)