@rowanwins 2017-04-25
Oh and it looks like the CI tests are failing because of a node precision thing in the tests. To get them to pass locally I had the change the values, however running in CI they seem to fail, grrr!