@DenisCarriere 2017-04-25
Odd stuff, I'll look into dropping those decimals.