@DenisCarriere 2017-04-25
@rowanwins I've been wanted to do this https://github.com/Turfjs/turf/pull/689/commits/b02d6adbaea5eaf23354738554ba9bbbe17a2a7f for a long time, `@turf/point-on-line` always had to manually edit the `expectedLocation` values because of decimals precision (not the module's fault).