@DenisCarriere 2017-04-25
@rowanwins Doesn't look like there was any performance loss from adding the `@turf/line-intersect` module.

## Benchmark Results

```
==after (@turf/line-intersect)==
line1 x 234,231 ops/sec ±1.78% (88 runs sampled)
route1 x 161 ops/sec ±1.53% (80 runs sampled)
route2 x 184 ops/sec ±1.96% (80 runs sampled)

==before (original)==
line1 x 272,785 ops/sec ±1.29% (87 runs sampled)
route1 x 195 ops/sec ±2.23% (80 runs sampled)
route2 x 218 ops/sec ±2.42% (78 runs sampled)
```

Those roads are REALLY long! I'm not surprised with these results.